Feat: 장바구니 담기 기능 및 회원탈퇴시 db 및 레디스 초기화 기능 구현

+++ src/main/java/shop/s5g/front/controller/CartController.java
@@ -4,11 +4,11 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.util.Map;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
-import org.springframework.boot.Banner.Mode;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -22,7 +22,9 @@ import shop.s5g.front.dto.cart.request.CartLoginRequestDto;
 import shop.s5g.front.dto.cart.request.CartPutRequestDto;
 import shop.s5g.front.dto.cart.request.CartRemoveBookRequestDto;
 import shop.s5g.front.dto.cart.request.CartUpdateQuantityRequestDto;
+import shop.s5g.front.exception.cart.CartConvertException;
 import shop.s5g.front.exception.cart.CartDetailPageException;
+import shop.s5g.front.exception.cart.CartPutException;
 import shop.s5g.front.service.cart.CartService;
 
 @Controller
@@ -31,11 +33,6 @@ public class CartController {
 
 
     private final CartService cartService;
-
-    @GetMapping("/cart/loginProcess")
-    public String cartLoginProcessing() {
-        return "cart/cartLogin";
-    }
 
     @GetMapping("/cart/loginCheck")
     public ResponseEntity<Map<String,Boolean>> cartLoginCheck(@CookieValue(value = "accessJwt",required = false) String isLoggedIn) {
@@ -48,26 +45,51 @@ public class CartController {
 
     // 로그인할 때 세션스토리지에 저장되었던 물품들을 레디스로 옮기고 회원 로그인상태에서는 레디스를 이용
     @PostMapping("/cart/login")
-    public ResponseEntity<Map<String,Integer>> convertCartToRedis(@RequestBody @Validated CartLoginRequestDto cartLoginRequestDto) {
-        Map<String, Integer> cartCount = cartService.convertCartToRedis(cartLoginRequestDto);
+    public ResponseEntity<Map<String, Integer>> convertCartToRedis(
+        @RequestBody @Validated CartLoginRequestDto cartLoginRequestDto,
+        BindingResult bindingResult) {
 
-        return ResponseEntity.status(HttpStatus.OK).body(cartCount);
+        if (bindingResult.hasErrors()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+
+        try {
+            Map<String, Integer> cartCount = cartService.convertCartToRedis(cartLoginRequestDto);
+
+            return ResponseEntity.status(HttpStatus.OK).body(cartCount);
+
+        } catch (CartConvertException e){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+
     }
-
+    //TODO 도서 상세페이지에 스크립트 작성할 때 예외 메시지를 정하는 걸로
     @PostMapping("/cart")
-    public String putBook(@RequestBody @Validated CartPutRequestDto cartPutRequestDto) {
+    public ResponseEntity<Map<String,String>> putBook(
+        @RequestBody @Validated CartPutRequestDto cartPutRequestDto, BindingResult bindingResult) {
 
-        MessageDto messageDto = cartService.putBook(cartPutRequestDto);
+        if (bindingResult.hasErrors()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
 
-        return messageDto.message();
+        try {
+            MessageDto messageDto = cartService.putBook(cartPutRequestDto);
+
+            return ResponseEntity.status(HttpStatus.OK)
+                .body(Map.of("message", messageDto.message()));
+
+        } catch (CartPutException e){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
     }
-
+    //TODO index.html 에서 alert 띄울듯
     @GetMapping("/cart/detailPage")
     public String cartDetailPage(Model model, RedirectAttributes redirectAttributes,
         @RequestParam(value = "cartBookInfoList", required = false) String cartSessionStorage,
-        HttpServletResponse response) {
+        HttpServletResponse response,
+        @CookieValue(value = "accessJwt", required = false) String isLoggedIn) {
 
-        if (cartSessionStorage == null) {
+        if (Objects.nonNull(isLoggedIn)) {
             // 로그인한 경우
             try {
                 cartDetailPageWhenLogin(response, model);
@@ -75,18 +97,21 @@ public class CartController {
                 return "cart/cartDetail";
 
             } catch (CartDetailPageException e) {
-                redirectAttributes.addFlashAttribute("errorMessage", e.getMessage());
+                redirectAttributes.addFlashAttribute("errorMessage", true);
 
                 return "redirect:/";
             }
         } else {
             // 로그인 하지 않은 경우
             try {
+                if (cartSessionStorage == null){
+                    return "cart/cartBackWhenLogout";
+                }
                 cartDetailPageWhenGuest(cartSessionStorage, response, model);
                 return "cart/cartDetail";
 
             } catch (CartDetailPageException e) {
-                redirectAttributes.addFlashAttribute("errorMessage", e.getMessage());
+                redirectAttributes.addFlashAttribute("errorMessage", true);
 
                 return "redirect:/";
             }
@@ -94,29 +119,43 @@ public class CartController {
     }
 
     @PostMapping("/cart/updateQuantity")
-    public ResponseEntity<?> updateQuantity(@RequestBody CartUpdateQuantityRequestDto cartUpdateQuantityReq) {
+    public ResponseEntity<Void> updateQuantity(
+        @RequestBody CartUpdateQuantityRequestDto cartUpdateQuantityReq,
+        BindingResult bindingResult) {
+
+        if (bindingResult.hasErrors()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+
         try {
             cartService.updateQuantity(cartUpdateQuantityReq);
 
             return ResponseEntity.status(HttpStatus.OK).build();
         } catch (CartDetailPageException e){
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                .body(Map.of("error", e.getMessage()));
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         }
     }
 
     @DeleteMapping("/cart/removeBook")
-    public ResponseEntity<?> removeBook(@RequestBody CartRemoveBookRequestDto cartRemoveBookReq) {
+    public ResponseEntity<Void> removeBook(@RequestBody CartRemoveBookRequestDto cartRemoveBookReq,
+        BindingResult bindingResult) {
+
+        if (bindingResult.hasErrors()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+
         try {
             cartService.removeBook(cartRemoveBookReq);
 
             return ResponseEntity.status(HttpStatus.OK).build();
-        }catch (CartDetailPageException e){
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                .body(Map.of("error", e.getMessage()));
+
+        } catch (CartDetailPageException e){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         }
 
     }
+
+
 
     private void cartDetailPageWhenLogin(HttpServletResponse response, Model model) {
         Map<String, Object> cartDetailPageInfo = cartService.getCartDetailPageInfo(); // 로그인 상태라면 api 에 가서 장바구니 정보를 가져와야한다.
